+++ src/app/(authenticated)/interview/page.tsx
@@ -1,5 +1,5 @@
-import { InterviewClient } from '@/featured/interview/components/InterviewClient'
+import { InterviewLayout } from '@/featured/interview/components/InterviewLayout'
 
 export default function InterviewPage() {
-  return <InterviewClient />
+  return <InterviewLayout />
 }
